java: move maven subgenerators to java or java-simple-application generators
