+++ generators/generate-blueprint/internal/lookup-namespaces.spec.ts
@@ -96,9 +96,6 @@ describe('lookupGeneratorsNamespaces', () => {
   "languages:bootstrap",
   "liquibase",
   "maven",
-  "maven:code-quality",
-  "maven:frontend-plugin",
-  "maven:jib",
   "project-name",
   "project-name:bootstrap",
   "react",
