+++ generators/java-simple-application/generators/code-quality/generator.spec.ts
@@ -52,7 +52,6 @@ describe(`generator - ${generator}`, () => {
       expect(result.getComposedGenerators()).toMatchInlineSnapshot(`
 [
   "jhipster:maven",
-  "jhipster:maven:code-quality",
 ]
 `);
     });
